created working tests
